# controlplane/gen/.openapi-3.0.yaml
openapi: 3.0.3
info:
  title: LangSmith Deployment
  version: 0.1.0
  description: "LangGraph Cloud-compatible API surface implemented by duragraph.\n\
    \nNote: paths are documented as bare (`/assistants`, `/threads`, `/runs`,\netc.)\
    \ to mirror the upstream LangGraph spec. The duragraph engine\nserves these paths\
    \ under the `/api/v1` base \u2014 declared via the\n`servers:` block below. Clients\
    \ should construct request URLs as\n`<server.url><path>`, e.g. `http://localhost:8081/api/v1/runs`.\n"
servers:
- url: http://localhost:8081/api/v1
  description: Local development server
- url: https://api.duragraph.ai/api/v1
  description: "Production server (placeholder \u2014 replace with deployment host)"
...
paths:
...
  /threads:
    post:
      tags:
      - Threads
      summary: Create Thread
      description: Create a thread.
      operationId: create_thread_threads_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThreadCreate'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thread'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /threads/search:
    post:
      tags:
      - Threads
      summary: Search Threads
      description: 'Search for threads.


        This endpoint also functions as the endpoint to list all threads.'
      operationId: search_threads_threads_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThreadSearchRequest'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Thread'
                type: array
                title: Response Search Threads Threads Search Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /threads/count:
    post:
      tags:
      - Threads
      summary: Count Threads
      description: Get the count of threads matching the specified criteria.
      operationId: count_threads_threads_count_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThreadCountRequest'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: integer
                title: Count
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /threads/{thread_id}/state:
    get:
      tags:
      - Threads
      summary: Get Thread State
      description: 'Get state for a thread.


        The latest state of the thread (i.e. latest checkpoint) is returned.'
      operationId: get_latest_thread_state_threads__thread_id__state_get
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      - description: Whether to include subgraphs in the response.
        required: false
        schema:
          type: boolean
          title: Subgraphs
          description: Whether to include subgraphs in the response.
        name: subgraphs
        in: query
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadState'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - Threads
      summary: Update Thread State
      description: Add state to a thread.
      operationId: update_thread_state_threads__thread_id__state_post
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThreadStateUpdate'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadStateUpdateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /threads/{thread_id}/state/{checkpoint_id}:
    get:
      tags:
      - Threads
      summary: Get Thread State At Checkpoint
      description: Get state for a thread at a specific checkpoint.
      operationId: get_thread_state_at_checkpoint_threads__thread_id__state__checkpoint_id__get
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      - description: The ID of the checkpoint.
        required: true
        schema:
          type: string
          format: uuid
          title: Checkpoint Id
          description: The ID of the checkpoint.
        name: checkpoint_id
        in: path
      - description: Whether to include subgraphs in the response.
        required: false
        schema:
          type: boolean
          title: Subgraphs
          description: Whether to include subgraphs in the response.
        name: subgraphs
        in: query
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadState'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /threads/{thread_id}/state/checkpoint:
    post:
      tags:
      - Threads
      summary: Get Thread State At Checkpoint
      description: Get state for a thread at a specific checkpoint.
      operationId: post_thread_state_at_checkpoint_threads__thread_id__state__checkpoint_id__get
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      - required: false
        schema:
          type: boolean
          title: Subgraphs
          description: If true, includes subgraph states.
        name: subgraphs
        in: query
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThreadStateCheckpointRequest'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadState'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /threads/{thread_id}/history:
    get:
      tags:
      - Threads
      summary: Get Thread History
      description: Get all past states for a thread.
      operationId: get_thread_history_threads__thread_id__history_get
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      - required: false
        schema:
          type: integer
          title: Limit
          default: 10
        name: limit
        in: query
      - required: false
        schema:
          type: string
          title: Before
        name: before
        in: query
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ThreadState'
                type: array
                title: Response Get Thread History Threads  Thread Id  History Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - Threads
      summary: Get Thread History Post
      description: Get all past states for a thread.
      operationId: get_thread_history_post_threads__thread_id__history_post
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThreadStateSearch'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ThreadState'
                type: array
                title: Response Get Thread History Post Threads  Thread Id  History
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /threads/{thread_id}/copy:
    post:
      tags:
      - Threads
      summary: Copy Thread
      description: Create a new thread with a copy of the state and checkpoints from
        an existing thread.
      operationId: copy_thread_post_threads__thread_id__copy_post
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thread'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /threads/{thread_id}:
    get:
      tags:
      - Threads
      summary: Get Thread
      description: Get a thread by ID.
      operationId: get_thread_threads__thread_id__get
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thread'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - Threads
      summary: Delete Thread
      description: Delete a thread by ID.
      operationId: delete_thread_threads__thread_id__delete
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema: {}
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - Threads
      summary: Patch Thread
      description: Update a thread.
      operationId: patch_thread_threads__thread_id__patch
      parameters:
      - description: The ID of the thread.
        required: true
        schema:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        name: thread_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThreadPatch'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thread'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ThreadSearchRequest:
      properties:
        ids:
          type: array
          items:
            type: string
            format: uuid
          title: Ids
          description: List of thread IDs to include. Others are excluded.
        metadata:
          type: object
          title: Metadata
          description: Thread metadata to filter on.
        values:
          type: object
          title: Values
          description: State values to filter on.
        status:
          type: string
          enum:
          - idle
          - busy
          - interrupted
          - error
          title: Status
          description: Thread status to filter on.
        limit:
          type: integer
          title: Limit
          description: Maximum number to return.
          default: 10
          minimum: 1
          maximum: 1000
        offset:
          type: integer
          title: Offset
          description: Offset to start from.
          default: 0
          minimum: 0
        sort_by:
          type: string
          enum:
          - thread_id
          - status
          - created_at
          - updated_at
          title: Sort By
          description: Sort by field.
        sort_order:
          type: string
          enum:
          - asc
          - desc
          title: Sort Order
          description: Sort order.
        select:
          type: array
          items:
            type: string
            enum:
            - thread_id
            - created_at
            - updated_at
            - metadata
            - config
            - context
            - status
            - values
            - interrupts
          title: Select
          description: Specify which fields to return. If not provided, all fields
            are returned.
      type: object
      title: ThreadSearchRequest
      description: Payload for listing threads.
    ThreadCountRequest:
      properties:
        metadata:
          type: object
          title: Metadata
          description: Thread metadata to filter on.
        values:
          type: object
          title: Values
          description: State values to filter on.
        status:
          type: string
          enum:
          - idle
          - busy
          - interrupted
          - error
          title: Status
          description: Thread status to filter on.
      type: object
      title: ThreadCountRequest
      description: Payload for counting threads.
    Thread:
      properties:
        thread_id:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread.
        created_at:
          type: string
          format: date-time
          title: Created At
          description: The time the thread was created.
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: The last time the thread was updated.
        metadata:
          type: object
          title: Metadata
          description: The thread metadata.
        config:
          type: object
          title: Config
          description: The thread config.
        status:
          type: string
          enum:
          - idle
          - busy
          - interrupted
          - error
          title: Status
          description: The status of the thread.
        values:
          type: object
          title: Values
          description: The current state of the thread.
        interrupts:
          type: object
          title: Interrupts
          description: The current interrupts of the thread.
      type: object
      required:
      - thread_id
      - created_at
      - updated_at
      - metadata
      - status
      title: Thread
    ThreadCreate:
      properties:
        thread_id:
          type: string
          format: uuid
          title: Thread Id
          description: The ID of the thread. If not provided, a random UUID will be
            generated.
        metadata:
          type: object
          title: Metadata
          description: Metadata to add to thread.
        if_exists:
          type: string
          enum:
          - raise
          - do_nothing
          title: If Exists
          description: How to handle duplicate creation. Must be either 'raise' (raise
            error if duplicate), or 'do_nothing' (return existing thread).
          default: raise
        ttl:
          type: object
          title: TTL
          description: The time-to-live for the thread.
          properties:
            strategy:
              type: string
              enum:
              - delete
              description: The TTL strategy. 'delete' removes the entire thread.
              default: delete
            ttl:
              type: number
              description: The time-to-live in minutes from now until thread should
                be swept.
        supersteps:
          type: array
          items:
            type: object
            properties:
              updates:
                type: array
                items:
                  $ref: '#/components/schemas/ThreadSuperstepUpdate'
            required:
            - updates
      type: object
      title: ThreadCreate
      description: Payload for creating a thread.
    ThreadPatch:
      properties:
        metadata:
          type: object
          title: Metadata
          description: Metadata to merge with existing thread metadata.
        ttl:
          type: object
          title: TTL
          description: The time-to-live for the thread.
          properties:
            strategy:
              type: string
              enum:
              - delete
              description: The TTL strategy. 'delete' removes the entire thread.
              default: delete
            ttl:
              type: number
              description: The time-to-live in minutes from now until thread should
                be swept.
      type: object
      title: ThreadPatch
      description: Payload for updating a thread.
    ThreadStateCheckpointRequest:
      properties:
        checkpoint:
          $ref: '#/components/schemas/CheckpointConfig'
          title: Checkpoint
          description: The checkpoint to get the state for.
        subgraphs:
          type: boolean
          title: Subgraphs
          description: Include subgraph states.
      required:
      - checkpoint
      type: object
      title: ThreadStateCheckpointRequest
      description: Payload for getting the state of a thread at a checkpoint.
    ThreadState:
      properties:
        values:
          title: Values
        next:
          items:
            type: string
          type: array
          title: Next
        tasks:
          items:
            type: object
            properties:
              id:
                type: string
                title: Task Id
              name:
                type: string
                title: Node Name
              error:
                type: string
                title: Error
              interrupts:
                type: array
                items:
                  $ref: '#/components/schemas/Interrupt'
              checkpoint:
                $ref: '#/components/schemas/CheckpointConfig'
                title: Checkpoint
              state:
                $ref: '#/components/schemas/ThreadState'
            required:
            - id
            - name
          type: array
          title: Tasks
        checkpoint:
          $ref: '#/components/schemas/CheckpointConfig'
          title: Checkpoint
        metadata:
          type: object
          title: Metadata
        created_at:
          type: string
          title: Created At
        parent_checkpoint:
          type: object
          title: Parent Checkpoint
        interrupts:
          type: array
          items:
            $ref: '#/components/schemas/Interrupt'
      type: object
      required:
      - values
      - next
      - checkpoint
      - metadata
      - created_at
      title: ThreadState
    ThreadStateSearch:
      properties:
        limit:
          type: integer
          title: Limit
          description: The maximum number of states to return.
          default: 1
          maximum: 1000
          minimum: 1
        before:
          title: Before
          description: Return states before this checkpoint.
          $ref: '#/components/schemas/CheckpointConfig'
        metadata:
          type: object
          title: Metadata
          description: Filter states by metadata key-value pairs.
        checkpoint:
          $ref: '#/components/schemas/CheckpointConfig'
          title: Checkpoint
          description: Return states for this subgraph.
      type: object
      title: ThreadStateSearch
    ThreadStateUpdate:
      properties:
        values:
          title: Values
          description: The values to update the state with.
        checkpoint:
          $ref: '#/components/schemas/CheckpointConfig'
          title: Checkpoint
          description: The checkpoint to update the state of.
        as_node:
          type: string
          title: As Node
          description: Update the state as if this node had just executed.
      type: object
      title: ThreadStateUpdate
      description: Payload for updating the state of a thread.
    ThreadSuperstepUpdate:
      properties:
        values: {}
        command:
          description: The command associated with the update.
        as_node:
          type: string
          description: Update the state as if this node had just executed.
      required:
      - as_node
      type: object
    ThreadStateUpdateResponse:
      properties:
        checkpoint:
          type: object
          title: Checkpoint
      type: object
      title: ThreadStateUpdateResponse
      description: Response for adding state to a thread.
    CheckpointConfig:
      type: object
      title: CheckpointConfig
      description: Checkpoint config.
      properties:
        thread_id:
          type: string
          description: Unique identifier for the thread associated with this checkpoint.
        checkpoint_ns:
          type: string
          description: Namespace for the checkpoint, used for organization and retrieval.
        checkpoint_id:
          type: string
          description: Optional unique identifier for the checkpoint itself.
        checkpoint_map:
          type: object
          description: Optional dictionary containing checkpoint-specific data.
...
    ErrorResponse:
      type: string
      title: ErrorResponse
      description: Error message returned from the server
    Interrupt:
      type: object
      title: Interrupt
      properties:
        id:
          type: string
          nullable: true
        value:
          type: object
      required:
      - value
...